Improved tests

+++ test/halfvec_test.cpp
@@ -8,12 +8,12 @@ using pgvector::HalfVector;
 
 static void test_constructor_vector() {
     auto vec = HalfVector({1, 2, 3});
-    assert_equal(vec.dimensions(), static_cast<size_t>(3));
+    assert_equal(vec.dimensions(), 3u);
 }
 
 static void test_constructor_span() {
     auto vec = HalfVector(std::span<const float>({1, 2, 3}));
-    assert_equal(vec.dimensions(), static_cast<size_t>(3));
+    assert_equal(vec.dimensions(), 3u);
 }
 
 void test_halfvec() {
